fix: resolve /wallet prerender crash (Turbopack SSR)
Two root causes fixed:

1. WithdrawalHistory.tsx: replace invalid useState(sideEffect) initializer
   with useEffect(() => fetchWithdrawalHistory(), []) — the useState misuse
   caused async code to run during SSR module evaluation, triggering the
   'Cannot access bk before initialization' circular-dep crash in Turbopack.

2. app/wallet/page.tsx: route segment config (export const dynamic) must
   live in a Server Component. Split into:
   - page.tsx (server): owns 'export const dynamic = force-dynamic'
   - WalletClient.tsx (client): all existing hooks/UI logic unchanged
